Update fetchReferenceContent() to return object instead of string
This object is a BibleReference object with a populated 'content'
property containing the textual contents of the reference. This is a
much more flexible return value, since it provides much more
information.

This, of course, is a breaking change, as is the removal of the
content-formatting logic. The consuming application is now responsible
for deciding how they format the reference content.