Fix coverage reporting under Vitest
Originally, in a03cc93, we updated the build pipeline to prebuild the
library before running the tests, then update the tests to import all
library modules/functions from the compiled project under dist/.

Part of the rationale for this was to make the tests behave more
realistically, as well as work with compile-time library integrations
(e.g. via @rollup/plugin-dynamic-import-vars). However, because Vitest
is ESM-based and uses Rollup under the hood, the dynamic imports are
automatically resolved (which would not be the case with Mocha or even
AVA if we were still importing from src/). 10 points to Vitest!