Add error message discrimination back to tests
These were taken out when we migrated to Vitest to solve some errors,
but now that we are handling the promises correctly with the
toThrowError() matcher, we can (and should) check against the specific
errors that are thrown. AVA had a nice instanceOf option to assert that
an instance of the specific error class was thrown. Jest/Vitest have no
such method, so we must resort to string-matching against the error
message.