Properly externalize htmlrewriter and html-entities
This solves a major issue I was hitting with bundling htmlrewriter with
my project. The issue was twofold:
- @rollup/plugin-wasm was breaking the initialization logic for
  htmlrewriter's WASM binary, thus preventing the WASM file from being
  loaded from the library bundle's entry point (dist/index.js).
 - htmlrewriter had its own set of exports to allow a consuming
   application's build tool to bundle the correct WASM initialization
   code according to the specific environment or runtime

By marking htmlrewriter as an external dependency in my Rollup config, I
solve both of these problems, in addition to simplifying the build
process and bundle size.

In accordance with this change, I also needed to mark the html-entities
package as external, lest Rollup present me with a warning when building
the library. Fortunately, the html-entities package did not strictly
need to be bundled with the library, anyway.

Of course, all of this creates the stipulation that any consuming
application which targets a browser (e.g. a browser extension or web
app) must have a build step so it can build the library. I think this
should be a fine stipulation to make, since I can't see many use cases
where a browser-targeting application would need to use this library AND
avoid a build step.